Disable unused modules and remove unnecessary files
This patch disables unused modules throughout the project by removing them
from module or dependency tags in pom.xml files, and the like. The
main criteria was anything that the agent did not need to run based
off the 'thermostat-plugin-agent.xml files for the distribution
modules in the project.

The files for these modules are removed from the repository.

Reviewed-by: jerboaa, neugens
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2017-March/022487.html