Fixes for Gradle compatibility
This is an attempt to be more consistent with the use of the "@property@" Maven resource filter.

I found that resource files that used ${property.with.dot} as a substitution syntax didn't work the same way that they do in Maven (the period character has a special meaning in Gradle), so I replaced it with the syntax @property@ instead.  Then I found that although I could see the resource had been transformed properly in target/classes (this is done by the <resources> element in the pom file), it was appearing untransformed in the jar files (this is done by the maven-bundle-plugin).

The syntax is laid out in http://felix.apache.org/documentation/subprojects/apache-felix-maven-bundle-plugin-bnd.html#include-resource
Using the curly braces tells BND to perform filtering while copying, the string "{maven-resources}" is actually the default behaviour(copy from src/main/resources to jar and filter), and the string "{path1=path2}" is copy with filtering to a specific location.

Admittedly this isn't strictly necessary if the old syntax for the property files is unchanged.  I wanted to make everything consistent, so adding this extra line seemed the best way.  It's imperfect, in that the pom file must explicitly name any files that need maven-style '@' filtering.

So my added line just tells BND to copy the pre-filtered property file from target/classes/..... into the jar file instead of performing its own filtering.
(I've removed the curly braces so BND doesn't perform a second filtering)


Reviewed-by: sgehwolf
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2017-March/022591.html