Add synchronization to VmListenerBackend
The backends will have a race condition in a later patch when there are multiple
threads calling activate and deactive, of which this adds synchronization so the
async activation/deactivation do not have a race condition. The race condition
occurs when two threads enter the deactivation() block for unregistering; they
both trigger the code that removes the backend from a list when it should only
be removed once (which in turn causes an exception).

Reviewed-by: jerboaa, aazores
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2017-July/024255.html